+++ microapps/FeatureFormsApp/app/src/main/java/com/arcgismaps/toolkit/featureformsapp/screens/map/MapScreen.kt
@@ -35,19 +35,38 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.window.core.layout.WindowSizeClass
 import androidx.window.layout.WindowMetricsCalculator
+import com.arcgismaps.data.Feature
+import com.arcgismaps.data.ServiceFeatureTable
 import com.arcgismaps.toolkit.composablemap.ComposableMap
 import com.arcgismaps.toolkit.featureforms.EditingTransactionState
 import com.arcgismaps.toolkit.featureforms.FeatureForm
 import com.arcgismaps.toolkit.featureformsapp.R
 import com.arcgismaps.toolkit.featureformsapp.screens.bottomsheet.BottomSheetMaxWidth
 import com.arcgismaps.toolkit.featureformsapp.screens.bottomsheet.SheetExpansionHeight
-import com.arcgismaps.toolkit.featureformsapp.screens.bottomsheet.SheetValue
 import com.arcgismaps.toolkit.featureformsapp.screens.bottomsheet.SheetLayout
+import com.arcgismaps.toolkit.featureformsapp.screens.bottomsheet.SheetValue
 import com.arcgismaps.toolkit.featureformsapp.screens.bottomsheet.StandardBottomSheet
 import com.arcgismaps.toolkit.featureformsapp.screens.bottomsheet.rememberStandardBottomSheetState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
+
+internal suspend fun commitEdits(feature: Feature): Result<Unit> {
+    val serviceFeatureTable =
+        feature.featureTable as? ServiceFeatureTable ?: return Result.failure(
+            IllegalStateException("cannot save feature edit without a ServiceFeatureTable")
+        )
+    
+    val result = serviceFeatureTable.updateFeature(feature)
+        .map {
+            serviceFeatureTable.serviceGeodatabase?.applyEdits()
+                ?: throw IllegalStateException("cannot apply feature edit without a ServiceGeodatabase")
+            feature.refresh()
+            Unit
+        }
+    
+    return result
+}
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -73,20 +92,23 @@ fun MapScreen(mapViewModel: MapViewModel = hiltViewModel(), onBackPressed: () ->
                 },
                 onSave = {
                     scope.launch {
-                        mapViewModel.commitEdits(EditingTransactionState.NotEditing)
-                            .onFailure {
-                                Log.w(
-                                    "Forms",
-                                    "applying edits from feature form failed with ${it.message}"
-                                )
-                                launch(Dispatchers.Main) {
-                                    Toast.makeText(
-                                        context,
-                                        "applying edits from feature form failed with ${it.message}",
-                                        Toast.LENGTH_LONG
-                                    ).show()
+                        mapViewModel.featureForm.value?.let { featureForm ->
+                            commitEdits(featureForm.feature)
+                                .onFailure {
+                                    Log.w(
+                                        "Forms",
+                                        "applying edits from feature form failed with ${it.message}"
+                                    )
+                                    launch(Dispatchers.Main) {
+                                        Toast.makeText(
+                                            context,
+                                            "applying edits from feature form failed with ${it.message}",
+                                            Toast.LENGTH_LONG
+                                        ).show()
+                                    }
                                 }
-                            }
+                        }
+                        mapViewModel.setTransactionState(EditingTransactionState.NotEditing)
                     }
                 }) {
                 onBackPressed()
